drcachesim: fixed FIFO implementation returning -1
Previously, the FIFO implementation of replace_which_way would return
-1 if something went wrong with the counters. As the call sites for the
function do not check for -1, this would lead to the wrong cache set
being evicted and replaced! (With the previous fixes, the FIFO counters
should always be in a valid state, and this should never happen. Even
so, the code has been modified to mitigate this strange behaviour.)